==> code/espejo/past/01-las-vegas-explore.Rmd <==
---
title: "Las Vegas"
author: "Edie Espejo"
date: "3/31/2019"
output: html_document
---

## Libraries
```{r, message=FALSE, warning=FALSE}
library(jsonlite)
library(tibble)
library(dplyr)
library(tidyr)
library(stringr)
library(readr)
library(ggplot2)
```

## Cleaning up the dataset
Let's check out businesses in SIN CITY!

```{r, warning=FALSE, message=FALSE}
causal_set <- read_csv("../../../data/causal_set.csv")[,-1]
causal_lv <- causal_set %>% filter(city=="Las Vegas") %>% filter(state=="NV")
causal_lv <- causal_lv %>% select(-city, -state)
```


I am removing those with no address.
```{r}
no_addy <- which(sapply(causal_lv %>% select(address), nchar)<10)
# length(no_addy)  # 1293
causal_lv <- causal_lv[-no_addy,]
# nrow(causal_lv)   # 20057
```

## Get zip codes
```{r}
home_rental_median <- read_csv("../../../data/zillow/Zip_MedianRentalPrice_AllHomes.csv")
```

I am merging on `2017-08` because our data is from August 22, 2017.

```{r}
names(home_rental_median)[which(names(home_rental_median)=="2017-08")] <- "august_2017_median"
zillow <- home_rental_median %>% select(RegionName, august_2017_median)
zillow <- zillow %>% rename(postal_code=RegionName)

complete_df <- inner_join(causal_lv, zillow, by="postal_code") %>% filter(!is.na(neighborhood))
```


```{r}
complete_df <- complete_df %>% filter(grepl("Restaurants|Food", categories)) %>% filter(!grepl("Grocery", categories))
```


```{r}
write.csv(x=complete_df, file="../../data/las-vegas-yelp.csv")
```


```{r}
set.seed(252)
subset_1 <- complete_df[runif(1500, 1, nrow(complete_df)),]
write.csv(x=subset_1, file="../../data/subset-1.csv")
```


```{r}
set.seed(4495)
subset_2 <- complete_df[runif(1500, 1, nrow(complete_df)),]
write.csv(x=subset_2, file="../../data/subset-2.csv")
```


## What businesses are we looking at?
```{r}
lv_business_types <- data.frame(sort(table(unlist(strsplit(x=causal_lv$categories, split=";"))), decreasing=TRUE))
ggplot(head(lv_business_types, 20), aes(x=Var1, y=Freq)) + geom_bar(stat="identity") + xlab("") + theme(axis.text.x = element_text(angle = 40, hjust = 1)) + ggtitle("Las Vegas Business Types")
```

## Which neighborhoods?
```{r}
neighborhoods <- data.frame(sort(table(causal_lv$neighborhood), decreasing=TRUE))
ggplot(head(neighborhoods, 20), aes(x=Var1, y=Freq)) + geom_bar(stat="identity") + xlab("") + theme(axis.text.x = element_text(angle = 40, hjust = 1)) + ggtitle("Las Vegas Neighborhoods")
```

## How many percent of businesses closed?
4% closed. Isn't that a small amount? Interesting...
```{r}
complete_df_sum <- complete_df %>% group_by(neighborhood) %>% 
  summarize(proportion_open=mean(open_2019),
            average_review=mean(stars),
            n=n()) %>%
  arrange(-n)
complete_df_sum
```

```{r}
ggplot(complete_df_sum, aes(x=neighborhood, y=average_review)) + geom_bar(stat="identity") + theme(axis.text.x = element_text(angle = 40, hjust = 1)) + ggtitle("Las Vegas Neighborhoods")
```


```{r}
ggplot(complete_df, aes(x=review_count, y=stars)) + geom_point(alpha=0.2)
```

```{r}
ggplot(complete_df, aes(y=stars)) + geom_boxplot() + facet_wrap(~open_2019) + ggtitle("Boxplots of business ratings that were closed/open in 2019")
```

```{r}
complete_df
```

```{r}
ix <- which(complete_df$neighborhood=="The Strip")
complete_df[ix,] %>% summarize(mean_rating=mean(stars),
                               sd_rating=sd(stars),
                               mean_count=mean(review_count),
                               sd_count=sd(review_count))
```


```{r}
standardized <- complete_df[ix,] %>% mutate(stars_standard=(stars-3.290909)/0.7679392) %>% mutate(count_standard=(review_count-271.2182)/353.1347)
ggplot(standardized, aes(x=stars_standard)) + geom_histogram(binwidth=0.5, col="white", lwd=0.2)
ggplot(standardized, aes(x=count_standard)) + geom_histogram(binwidth=0.5, col="white", lwd=0.2)

```

```{r}
ggplot(complete_df, aes(y=review_count)) + geom_boxplot() + facet_wrap(~open_2019) + ggtitle("Boxplots of business reviews that were closed/open in 2019")
```
