==> code/espejo/past/02-food-filters.Rmd <==
---
title: "Playing with data"
author: "Asem Berkalieva"
date: "4/10/2019"
output: pdf_document
--- 

```{r}
library(tidyverse)

# read in data
vegas <- read.csv("../../data/las-vegas-yelp.csv")

# remove any other non-food places

vegas_updated <- vegas %>% filter(!grepl("Walgreens", name)) %>%
  filter(!grepl("Specialty Food", categories)) %>%
  filter(!grepl("Convenience Stores", categories)) %>%
  filter(!grepl("Gas Stations", categories)) %>%
  filter(!grepl("Car Wash", categories)) %>%
  filter(!grepl("Wine & Spirits", categories)) %>%
  filter(!grepl("Festival", categories)) %>%
  filter(!grepl("Water Store", categories))

write.csv(vegas_updated, file = "../../data/vegas-updated.csv")

```
